# fix: Address inadvertent quadratic behaviour in `expand_columns`